structured-llm-eval-framework/requirements.txt
fastapi
uvicorn
azure-ai-openai
pydantic
jsonschema
httpx
pytest
python-dotenv